backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "octopus-secondbrain-backend"
version = "2.0.0"
description = "Octopus AI Second Brain - Multi-Modal RAG Backend"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Octopus AI Second Brain Team"}
]
keywords = ["rag", "ai", "knowledge-management", "second-brain", "multimodal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core FastAPI stack
    "fastapi>=0.117.1",
    "uvicorn[standard]>=0.36.0",
    "gunicorn>=21.2.0",
    "python-multipart>=0.0.20",
    # Database & ORM
    "sqlalchemy>=2.0.43",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
    "alembic>=1.13.1",
    # Pydantic & Settings
    "pydantic>=2.11.9",
    "pydantic-settings>=2.2.1",
    # Security
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=41.0.0",
    # RAG dependencies
    # ML & Embeddings
    "numpy>=1.24.0,<2.0.0",
    "sentence-transformers>=3.0.0",
    "transformers>=4.45.0",
    "torch>=2.0.0",
    "Pillow>=10.0.0",
    "faiss-cpu>=1.8.0",
    # Optional RAG dependencies
    "PyPDF2>=3.0.0",
    "pytesseract>=0.3.13",
    "openai>=1.108.0",
    # Utilities
    "python-dotenv>=1.1.1",
    "slowapi>=0.1.9",
    "email-validator>=2.3.0",
    "python-magic>=0.4.27",
    "moviepy>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "black>=24.0.0",
    "mypy>=1.13.0",
    "pre-commit>=3.6.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
]

audio = [
    "soundfile>=0.12.1",
    "librosa>=0.10.0",
    "moviepy>=1.0.3",
]

chromadb = [
    "chromadb>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/Octopus-AI-SecondBrain/Octopus-AI-SecondBrain.github.io"
Repository = "https://github.com/Octopus-AI-SecondBrain/Octopus-AI-SecondBrain.github.io"
Issues = "https://github.com/Octopus-AI-SecondBrain/Octopus-AI-SecondBrain.github.io/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "sentence_transformers.*",
    "faiss.*",
    "chromadb.*",
    "PyPDF2.*",
    "pytesseract.*",
    "soundfile.*",
    "librosa.*",
    "moviepy.*",
    "transformers.*",
    "torch.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[dependency-groups]
dev = [
    "faker>=38.0.0",
    "redis>=7.0.1",
]
